+++ packages/toolboxes/toolbox-cosmos/src/binanceUtils/index.ts
@@ -1,2 +1,3 @@
+export { sortObject } from './amino/encoder.js';
 export * from './types.js';
 export * from './utils.js';