+++ packages/toolboxes/toolbox-evm/src/toolbox/arb.ts
@@ -1,33 +1,35 @@
 import { Provider } from '@ethersproject/abstract-provider';
 import { Signer } from '@ethersproject/abstract-signer';
-import { Web3Provider } from '@ethersproject/providers';
+import { JsonRpcProvider, Web3Provider } from '@ethersproject/providers';
 import { baseAmount } from '@thorswap-lib/helpers';
-import { AssetEntity, getSignatureAssetFor } from '@thorswap-lib/swapkit-entities';
-import { Address, BaseDecimal, Chain, ChainId } from '@thorswap-lib/types';
+import { getSignatureAssetFor } from '@thorswap-lib/swapkit-entities';
+import {
+  Address,
+  BaseDecimal,
+  Chain,
+  ChainId,
+  ChainToExplorerUrl,
+  FeeOption,
+  RPCUrl,
+} from '@thorswap-lib/types';
 
 import { covalentApi, CovalentApiType } from '../api/covalentApi.js';
 import { getProvider } from '../provider.js';
 
 import { BaseEVMToolbox } from './BaseEVMToolbox.js';
 
-export const getBalance = async (
-  api: CovalentApiType,
-  address: Address,
-  assets?: AssetEntity[],
-) => {
+export const getBalance = async (api: CovalentApiType, address: Address) => {
   const provider = getProvider(Chain.Arbitrum);
   const tokenBalances = await api.getBalance(address);
 
-  if (assets) {
-    return tokenBalances.filter(({ asset }) =>
-      assets.find(({ chain, symbol }) => chain === asset.chain && symbol === asset.symbol),
-    );
-  }
-
   const evmGasTokenBalance = await provider.getBalance(address);
   const evmGasTokenBalanceAmount = baseAmount(evmGasTokenBalance, BaseDecimal.ARB);
+
   return [
-    { asset: getSignatureAssetFor(Chain.Arbitrum), amount: evmGasTokenBalanceAmount },
+    {
+      asset: getSignatureAssetFor(Chain.Arbitrum),
+      amount: evmGasTokenBalanceAmount,
+    },
     ...tokenBalances,
   ];
 };
@@ -35,14 +37,28 @@ export const getBalance = async (
 export const getNetworkParams = () => ({
   chainId: ChainId.ArbitrumHex,
   chainName: 'Arbitrum One',
-  nativeCurrency: {
-    name: 'Arbitrum',
-    symbol: 'ARB',
-    decimals: 18,
-  },
-  rpcUrls: ['https://arb1.arbitrum.io/rpc'],
-  blockExplorerUrls: ['https://arbiscan.io/'],
+  nativeCurrency: { name: 'Ethereum', symbol: Chain.Ethereum, decimals: BaseDecimal.ETH },
+  rpcUrls: [RPCUrl.Arbitrum],
+  blockExplorerUrls: [ChainToExplorerUrl[Chain.Arbitrum]],
 });
+
+const estimateGasPrices = async (provider: Provider) => {
+  try {
+    const { gasPrice } = await provider.getFeeData();
+
+    if (!gasPrice) throw new Error('No fee data available');
+
+    return {
+      [FeeOption.Average]: { gasPrice },
+      [FeeOption.Fast]: { gasPrice },
+      [FeeOption.Fastest]: { gasPrice },
+    };
+  } catch (error) {
+    throw new Error(
+      `Failed to estimate gas price: ${(error as any).msg ?? (error as any).toString()}`,
+    );
+  }
+};
 
 export const ARBToolbox = ({
   api,
@@ -53,13 +69,15 @@ export const ARBToolbox = ({
   api?: CovalentApiType;
   covalentApiKey: string;
   signer: Signer;
-  provider: Provider | Web3Provider;
+  provider: JsonRpcProvider | Web3Provider;
 }) => {
   const arbApi = api || covalentApi({ apiKey: covalentApiKey, chainId: ChainId.Arbitrum });
+  const baseToolbox = BaseEVMToolbox({ provider, signer, isEIP1559Compatible: false });
 
   return {
-    ...BaseEVMToolbox({ provider, signer }),
+    ...baseToolbox,
     getNetworkParams,
-    getBalance: (address: string, assets?: AssetEntity[]) => getBalance(arbApi, address, assets),
+    estimateGasPrices: () => estimateGasPrices(provider),
+    getBalance: (address: string) => getBalance(arbApi, address),
   };
 };