+++ packages/toolboxes/toolbox-evm/src/toolbox/matic.ts
@@ -1,8 +1,7 @@
-import { Provider } from '@ethersproject/abstract-provider';
 import { Signer } from '@ethersproject/abstract-signer';
-import { Web3Provider } from '@ethersproject/providers';
+import { JsonRpcProvider, Web3Provider } from '@ethersproject/providers';
 import { baseAmount } from '@thorswap-lib/helpers';
-import { AssetEntity, getSignatureAssetFor } from '@thorswap-lib/swapkit-entities';
+import { getSignatureAssetFor } from '@thorswap-lib/swapkit-entities';
 import {
   Address,
   BaseDecimal,
@@ -17,22 +16,12 @@ import { getProvider } from '../provider.js';
 
 import { BaseEVMToolbox } from './BaseEVMToolbox.js';
 
-export const getBalance = async (
-  api: CovalentApiType,
-  address: Address,
-  assets?: AssetEntity[],
-) => {
+export const getBalance = async (api: CovalentApiType, address: Address) => {
   const provider = getProvider(Chain.Polygon);
   const tokenBalances = await api.getBalance(address);
-
-  if (assets) {
-    return tokenBalances.filter(({ asset }) =>
-      assets.find(({ chain, symbol }) => chain === asset.chain && symbol === asset.symbol),
-    );
-  }
-
   const evmGasTokenBalance = await provider.getBalance(address);
   const evmGasTokenBalanceAmount = baseAmount(evmGasTokenBalance, BaseDecimal.MATIC);
+
   return [
     { asset: getSignatureAssetFor(Chain.Polygon), amount: evmGasTokenBalanceAmount },
     ...tokenBalances,
@@ -41,12 +30,8 @@ export const getBalance = async (
 
 export const getNetworkParams = () => ({
   chainId: ChainId.PolygonHex,
-  chainName: 'Polygon',
-  nativeCurrency: {
-    name: 'Polygon',
-    symbol: Chain.Polygon,
-    decimals: 18,
-  },
+  chainName: 'Polygon Mainnet',
+  nativeCurrency: { name: 'Polygon', symbol: Chain.Polygon, decimals: BaseDecimal.MATIC },
   rpcUrls: [RPCUrl.Polygon],
   blockExplorerUrls: [ChainToExplorerUrl[Chain.Polygon]],
 });
@@ -60,13 +45,14 @@ export const MATICToolbox = ({
   api?: CovalentApiType;
   covalentApiKey: string;
   signer: Signer;
-  provider: Provider | Web3Provider;
+  provider: JsonRpcProvider | Web3Provider;
 }) => {
   const maticApi = api || covalentApi({ apiKey: covalentApiKey, chainId: ChainId.Polygon });
+  const baseToolbox = BaseEVMToolbox({ provider, signer });
 
   return {
-    ...BaseEVMToolbox({ provider, signer }),
+    ...baseToolbox,
     getNetworkParams,
-    getBalance: (address: string, assets?: AssetEntity[]) => getBalance(maticApi, address, assets),
+    getBalance: (address: string) => getBalance(maticApi, address),
   };
 };