+++ packages/toolboxes/toolbox-evm/src/toolbox/op.ts
@@ -1,38 +1,84 @@
-import { Provider } from '@ethersproject/abstract-provider';
+import { Provider, TransactionRequest } from '@ethersproject/abstract-provider';
 import { Signer } from '@ethersproject/abstract-signer';
-import { Web3Provider } from '@ethersproject/providers';
+import { BigNumber } from '@ethersproject/bignumber';
+import { Contract } from '@ethersproject/contracts';
+import { JsonRpcProvider, Web3Provider } from '@ethersproject/providers';
+import { serialize } from '@ethersproject/transactions';
 import { baseAmount } from '@thorswap-lib/helpers';
-import { AssetEntity, getSignatureAssetFor } from '@thorswap-lib/swapkit-entities';
+import { getSignatureAssetFor } from '@thorswap-lib/swapkit-entities';
 import {
   Address,
   BaseDecimal,
   Chain,
   ChainId,
   ChainToExplorerUrl,
+  FeeOption,
   RPCUrl,
 } from '@thorswap-lib/types';
 
 import { covalentApi, CovalentApiType } from '../api/covalentApi.js';
+import { gasOracleAbi } from '../contracts/op/gasOracle.js';
 import { getProvider } from '../provider.js';
 
 import { BaseEVMToolbox } from './BaseEVMToolbox.js';
 
-export const getBalance = async (
-  api: CovalentApiType,
-  address: Address,
-  assets?: AssetEntity[],
+const GAS_PRICE_ORACLE_ADDRESS = '0x420000000000000000000000000000000000000f';
+
+export const connectGasPriceOracle = (provider: Provider) => {
+  return new Contract(GAS_PRICE_ORACLE_ADDRESS, gasOracleAbi, provider);
+};
+
+export const getL1GasPrice = async (provider: Provider) => {
+  return connectGasPriceOracle(provider).l1BaseFee();
+};
+
+const _serializeTx = async (
+  provider: Provider,
+  { data, from, to, gasPrice, type, gasLimit, nonce }: TransactionRequest,
 ) => {
+  return serialize({
+    data,
+    to,
+    gasPrice,
+    type,
+    gasLimit,
+    nonce: nonce
+      ? BigNumber.from(nonce).toNumber()
+      : from
+      ? await provider.getTransactionCount(from)
+      : 0,
+  });
+};
+
+export const estimateL1GasCost = async (provider: Provider, tx: TransactionRequest) => {
+  return connectGasPriceOracle(provider).getL1Fee(await _serializeTx(provider, tx));
+};
+
+export const estimateL2GasCost = async (
+  provider: Provider,
+  tx: TransactionRequest,
+): Promise<BigNumber> => {
+  const l2GasPrice = await provider.getGasPrice();
+  const l2GasCost = await provider.estimateGas(tx);
+  return l2GasPrice.mul(l2GasCost);
+};
+
+export const estimateTotalGasCost = async (provider: Provider, tx: TransactionRequest) => {
+  const l1GasCost = await estimateL1GasCost(provider, tx);
+  const l2GasCost = await estimateL2GasCost(provider, tx);
+  return l1GasCost.add(l2GasCost);
+};
+
+export const estimateL1Gas = async (provider: Provider, tx: TransactionRequest) => {
+  return connectGasPriceOracle(provider).getL1GasUsed(await _serializeTx(provider, tx));
+};
+
+export const getBalance = async (api: CovalentApiType, address: Address) => {
   const provider = getProvider(Chain.Optimism);
   const tokenBalances = await api.getBalance(address);
-
-  if (assets) {
-    return tokenBalances.filter(({ asset }) =>
-      assets.find(({ chain, symbol }) => chain === asset.chain && symbol === asset.symbol),
-    );
-  }
-
   const evmGasTokenBalance = await provider.getBalance(address);
   const evmGasTokenBalanceAmount = baseAmount(evmGasTokenBalance, BaseDecimal.OP);
+
   return [
     { asset: getSignatureAssetFor(Chain.Optimism), amount: evmGasTokenBalanceAmount },
     ...tokenBalances,
@@ -42,14 +88,44 @@ export const getBalance = async (
 export const getNetworkParams = () => ({
   chainId: ChainId.OptimismHex,
   chainName: 'Optimism',
-  nativeCurrency: {
-    name: 'Optimism',
-    symbol: Chain.Optimism,
-    decimals: 18,
-  },
+  nativeCurrency: { name: 'Ethereum', symbol: Chain.Ethereum, decimals: BaseDecimal.ETH },
   rpcUrls: [RPCUrl.Optimism],
   blockExplorerUrls: [ChainToExplorerUrl[Chain.Optimism]],
 });
+
+const estimateGasPrices = async (provider: Provider) => {
+  try {
+    const { maxFeePerGas, maxPriorityFeePerGas, gasPrice } = await provider.getFeeData();
+    const l1GasPrice = await getL1GasPrice(provider);
+
+    if (!maxFeePerGas || !maxPriorityFeePerGas) throw new Error('No fee data available');
+
+    return {
+      [FeeOption.Average]: {
+        l1GasPrice,
+        gasPrice,
+        maxFeePerGas,
+        maxPriorityFeePerGas,
+      },
+      [FeeOption.Fast]: {
+        l1GasPrice: l1GasPrice.mul(15).div(10),
+        gasPrice: gasPrice?.mul(15).div(10),
+        maxFeePerGas,
+        maxPriorityFeePerGas: maxPriorityFeePerGas.mul(15).div(10),
+      },
+      [FeeOption.Fastest]: {
+        l1GasPrice: l1GasPrice.mul(2),
+        gasPrice: gasPrice?.mul(2),
+        maxFeePerGas,
+        maxPriorityFeePerGas: maxPriorityFeePerGas.mul(2),
+      },
+    };
+  } catch (error) {
+    throw new Error(
+      `Failed to estimate gas price: ${(error as any).msg ?? (error as any).toString()}`,
+    );
+  }
+};
 
 export const OPToolbox = ({
   api,
@@ -60,13 +136,20 @@ export const OPToolbox = ({
   api?: CovalentApiType;
   covalentApiKey: string;
   signer: Signer;
-  provider: Provider | Web3Provider;
+  provider: JsonRpcProvider | Web3Provider;
 }) => {
   const opApi = api || covalentApi({ apiKey: covalentApiKey, chainId: ChainId.Optimism });
+  const baseToolbox = BaseEVMToolbox({ provider, signer });
 
   return {
-    ...BaseEVMToolbox({ provider, signer }),
+    ...baseToolbox,
+    estimateTotalGasCost: (tx: TransactionRequest) => estimateTotalGasCost(provider, tx),
+    estimateL1GasCost: (tx: TransactionRequest) => estimateL1GasCost(provider, tx),
+    estimateL2GasCost: (tx: TransactionRequest) => estimateL2GasCost(provider, tx),
+    getL1GasPrice: () => getL1GasPrice(provider),
+    estimateL1Gas: (tx: TransactionRequest) => estimateL1Gas(provider, tx),
     getNetworkParams,
-    getBalance: (address: string, assets?: AssetEntity[]) => getBalance(opApi, address, assets),
+    estimateGasPrices: () => estimateGasPrices(provider),
+    getBalance: (address: string) => getBalance(opApi, address),
   };
 };