+++ packages/toolboxes/toolbox-evm/src/types/etherscan-api-types.ts
@@ -1,8 +1,0 @@
-export type FullBalanceParams = {
-  address: string;
-};
-
-export type TokenBalanceParam = {
-  address: string;
-  assetAddress: string;
-};