+++ packages/wallets/trustwallet/src/index.ts
@@ -1,4 +1,0 @@
-export * from './constants.js';
-export * from './trustwallet.js';
-export * from './types.js';
-export { default as WalletConnectProvider } from '@walletconnect/web3-provider';