+++ packages/wallets/walletconnect/src/walletconnect.ts
@@ -1,86 +1,254 @@
+import { makeSignDoc as makeSignDocAmino, StdSignDoc } from '@cosmjs/amino';
+import {
+  BinanceToolbox,
+  DepositParam,
+  getDenomWithChain,
+  sortObject,
+  ThorchainToolbox,
+} from '@thorswap-lib/toolbox-cosmos';
 import { AVAXToolbox, ETHToolbox, getProvider } from '@thorswap-lib/toolbox-evm';
-import { Chain, WalletOption } from '@thorswap-lib/types';
+import { Chain, ChainId, WalletOption, WalletTxParams } from '@thorswap-lib/types';
 import QRCodeModal from '@walletconnect/qrcode-modal';
-import Client, { type SignClient } from '@walletconnect/sign-client';
-import { SessionTypes, SignClientTypes } from '@walletconnect/types';
+import Client from '@walletconnect/sign-client';
+import type { SessionTypes, SignClientTypes } from '@walletconnect/types';
+import { auth, StdTx } from 'cosmos-client/x/auth/index.js';
 
 import {
+  BINANCE_MAINNET_ID,
   DEFAULT_APP_METADATA,
+  DEFAULT_COSMOS_METHODS,
   DEFAULT_LOGGER,
   DEFAULT_RELAY_URL,
+  THORCHAIN_MAINNET_ID,
   WC_SUPPORTED_CHAINS,
 } from './constants.js';
 import { getEVMSigner } from './evmSigner.js';
-import { chainToChainId } from './helpers.js';
+import { chainToChainId, getAddressByChain } from './helpers.js';
 import { getRequiredNamespaces } from './namespaces.js';
 
-// const getToolbox = async ({
-//   chain,
-//   ethplorerApiKey,
-//   covalentApiKey,
-//   address,
-//   walletconnectClient,
-//   session,
-// }: {
-//   // @ts-ignore
-//   walletconnectClient: Client;
-//   session: SessionTypes.Struct;
-//   chain: (typeof SUPPORTED_CHAINS)[number];
-//   covalentApiKey?: string;
-//   ethplorerApiKey?: string;
-//   address: string;
-// }) => {
-//   const from = address;
+const THORCHAIN_GAS_FEE = '500000000';
+const DEFAULT_THORCHAIN_FEE = {
+  amount: [],
+  gas: THORCHAIN_GAS_FEE,
+};
 
-//   switch (chain) {
-//     case Chain.Avalanche:
-//     case Chain.BinanceSmartChain:
-//     case Chain.Ethereum: {
-//       const provider = getProvider(chain);
+const SUPPORTED_CHAINS = [
+  Chain.Binance,
+  Chain.BinanceSmartChain,
+  Chain.Ethereum,
+  Chain.THORChain,
+  Chain.Avalanche,
+] as const;
 
-//       const signer = {
-//         getAddress: async () => address,
-//         _isSigner: true,
-//         sendTransaction: async (tx: EIP1559TxParams) => {
-//           const txHash: string = await walletconnectClient.request({
-//             chainId: chainToChainId(chain),
-//             topic: session.topic,
-//             request: {
-//               method: DEFAULT_EIP155_METHODS.ETH_SEND_TRANSACTION,
-//               params: [
-//                 {
-//                   from,
-//                   to: tx.to?.toLocaleLowerCase(),
-//                   data: tx.data || undefined,
-//                   value: BigNumber.from(tx.value || 0).toHexString(),
-//                 },
-//               ],
-//             },
-//           });
-//           return txHash;
-//         },
-//       } as unknown as Signer;
+const getToolbox = async ({
+  chain,
+  ethplorerApiKey,
+  covalentApiKey = '',
+  walletconnect,
+  address,
+  session,
+}: {
+  // @ts-ignore
+  walletconnect: Walletconnect;
+  session: SessionTypes.Struct;
+  chain: (typeof SUPPORTED_CHAINS)[number];
+  covalentApiKey?: string;
+  ethplorerApiKey?: string;
+  address: string;
+}) => {
+  const from = address;
 
-//       return toolbox;
-//     }
-//     default:
-//       throw new Error('Chain is not supported');
-//   }
-// };
+  switch (chain) {
+    case Chain.Avalanche:
+    case Chain.Ethereum: {
+      if (chain === Chain.Ethereum && !ethplorerApiKey)
+        throw new Error('Ethplorer API key not found');
+      if (chain !== Chain.Ethereum && !covalentApiKey)
+        throw new Error('Covalent API key not found');
+
+      const provider = getProvider(chain);
+      const signer = await getEVMSigner({ walletconnect, chain, provider });
+
+      const toolbox =
+        chain === Chain.Ethereum
+          ? ETHToolbox({ provider, signer, ethplorerApiKey: ethplorerApiKey as string })
+          : AVAXToolbox({ provider, signer, covalentApiKey });
+
+      return toolbox;
+    }
+    case Chain.Binance: {
+      const toolbox = BinanceToolbox();
+      const transfer = async (params: any) => {
+        const account = await toolbox.getAccount(from);
+        const txAmount = params.amount.amount().toString();
+        const { transaction, signMsg } = await toolbox.createTransactionAndSignMsg({
+          from,
+          to: params.recipient,
+          amount: txAmount,
+          asset: params.asset.ticker,
+          memo: params.memo,
+        });
+
+        const signDoc = sortObject({
+          account_number: account.account_number.toString(),
+          chain_id: ChainId.Binance,
+          data: null,
+          memo: params.memo,
+          msgs: [signMsg],
+          sequence: account.sequence.toString(),
+          source: '0',
+        });
+
+        const response: any = await walletconnect?.client.request({
+          chainId: BINANCE_MAINNET_ID,
+          topic: session.topic,
+          request: {
+            method: DEFAULT_COSMOS_METHODS.COSMOS_SIGN_AMINO,
+            params: { signerAddress: address, signDoc },
+          },
+        });
+
+        const signature = Buffer.from(response.signature, 'hex');
+        const publicKey = toolbox.getPublicKey(response.publicKey);
+        const signedTx = transaction.addSignature(publicKey, signature);
+
+        const res = await toolbox.sendRawTransaction(signedTx.serialize(), true);
+
+        return res[0]?.hash;
+      };
+      return { ...toolbox, transfer };
+    }
+    case Chain.THORChain: {
+      const toolbox = ThorchainToolbox({ stagenet: false });
+
+      const signRequest = (signDoc: StdSignDoc) =>
+        walletconnect?.client.request({
+          chainId: THORCHAIN_MAINNET_ID,
+          topic: session.topic,
+          request: {
+            method: DEFAULT_COSMOS_METHODS.COSMOS_SIGN_AMINO,
+            params: { signerAddress: address, signDoc },
+          },
+        });
+
+      const transfer = async (params: WalletTxParams) => {
+        const account = await toolbox.getAccount(from);
+        if (!account?.account_number) throw new Error('Account not found');
+        const { account_number: accountNumber, sequence = '0' } = account;
+
+        const sendCoinsMessage = {
+          amount: [
+            {
+              amount: params.amount.amount().toString(),
+              denom: params.asset?.symbol.toLowerCase(),
+            },
+          ],
+          from_address: address,
+          to_address: params.recipient,
+        };
+
+        const msg = {
+          type: 'thorchain/MsgSend',
+          value: sendCoinsMessage,
+        };
+
+        const signDoc = makeSignDocAmino(
+          [msg],
+          DEFAULT_THORCHAIN_FEE,
+          ChainId.THORChain,
+          params.memo,
+          accountNumber?.toString(),
+          sequence?.toString() || '0',
+        );
+
+        const signature: any = await signRequest(signDoc);
+
+        const txObj = {
+          msg: [msg],
+          fee: DEFAULT_THORCHAIN_FEE,
+          memo: params.memo,
+          signatures: [
+            {
+              ...signature,
+              sequence: sequence?.toString(),
+            },
+          ],
+        };
+
+        const stdTx = StdTx.fromJSON(txObj);
+
+        const res: any = await auth.txsPost(toolbox.sdk as any, stdTx, 'sync');
+
+        return res.data?.txhash;
+      };
+
+      const deposit = async ({ asset, amount, memo }: DepositParam) => {
+        const account = await toolbox.getAccount(address);
+        if (!asset) throw new Error('invalid asset to deposit');
+        if (!account?.account_number) throw new Error('Account not found');
+        const { account_number: accountNumber, sequence = '0' } = account;
+
+        const msg = {
+          type: 'thorchain/MsgDeposit',
+          value: {
+            coins: [
+              {
+                amount: amount.amount().toString(),
+                asset: getDenomWithChain(asset).toUpperCase(),
+              },
+            ],
+            memo,
+            signer: address,
+          },
+        };
+
+        const signDoc = makeSignDocAmino(
+          [msg],
+          DEFAULT_THORCHAIN_FEE,
+          ChainId.THORChain,
+          memo,
+          accountNumber?.toString(),
+          sequence?.toString() || '0',
+        );
+
+        const signature: any = await signRequest(signDoc);
+
+        const txObj = {
+          msg: [msg],
+          fee: DEFAULT_THORCHAIN_FEE,
+          memo,
+          signatures: [
+            {
+              ...signature,
+              sequence: sequence?.toString(),
+            },
+          ],
+        };
+
+        const stdTx = StdTx.fromJSON(txObj);
+
+        const res: any = await auth.txsPost(toolbox.sdk as any, stdTx, 'sync');
+
+        return res.data?.txhash;
+      };
+
+      return { ...toolbox, transfer, deposit };
+    }
+    default:
+      throw new Error('Chain is not supported');
+  }
+};
 
 const getWalletconnect = async (
   chains: Chain[],
   walletConnectProjectId?: string,
   walletconnectOptions?: SignClientTypes.Options,
-): Promise<{
-  session: SessionTypes.Struct;
-  accounts: string[];
-  client: InstanceType<typeof SignClient>;
-}> => {
+) => {
   try {
     const requiredNamespaces = getRequiredNamespaces(chains.map(chainToChainId));
-    //@ts-ignore
-    const client = await (Client as typeof Client.SignClient).init({
+
+    // @ts-ignore
+    const client = await Client.init({
       logger: DEFAULT_LOGGER,
       relayUrl: DEFAULT_RELAY_URL,
       projectId: walletConnectProjectId,
@@ -95,24 +263,24 @@ const getWalletconnect = async (
     // Open QRCode modal if a URI was returned (i.e. we're not connecting an existing pairing).
     if (uri) {
       // @ts-ignore
-      (QRCodeModal as typeof QRCodeModal.default).open(uri, () => {
+      QRCodeModal.open(uri, () => {
         console.log('EVENT', 'QR Code Modal closed');
       });
     }
 
     const session = await approval();
-    const accounts = Object.values<SessionTypes.Namespace>(session.namespaces)
-      .map((namespace) => namespace.accounts)
+    const accounts = Object.values(session.namespaces)
+      .map((namespace: any) => namespace.accounts)
       .flat();
 
-    //@ts-ignore
     return { session, accounts, client };
   } catch (e) {
-    throw new Error('Unable to establish connection through walletconnect');
+    console.error(e);
   } finally {
     // @ts-ignore
-    (QRCodeModal as typeof QRCodeModal.default).close();
+    QRCodeModal.close();
   }
+  return undefined;
 };
 
 export type Walletconnect = Awaited<ReturnType<typeof getWalletconnect>>;
@@ -140,46 +308,28 @@ const connectWalletconnect =
       walletconnectOptions,
     );
 
+    if (!walletconnect) throw new Error('Unable to establish connection through walletconnect');
+
+    const { session, accounts } = walletconnect;
+
     const promises = chainsToConnect.map(async (chain) => {
-      if (
-        chain === Chain.Avalanche ||
-        chain === Chain.BinanceSmartChain ||
-        chain === Chain.Ethereum
-      ) {
-        // TODO improve check if keys are set
-        if (chain === Chain.Ethereum && !ethplorerApiKey)
-          throw new Error('Ethplorer API key not found');
-        if (chain !== Chain.Ethereum && !covalentApiKey)
-          throw new Error('Covalent API key not found');
+      const address = getAddressByChain(chain, accounts);
 
-        // TODO(@0xGeneral): Check on typings
-        const provider = getProvider(chain) as unknown as any;
-        const signer = await getEVMSigner({ walletconnect, chain, provider });
-        const toolbox =
-          chain === Chain.Ethereum && ethplorerApiKey
-            ? ETHToolbox({
-                provider,
-                signer,
-                ethplorerApiKey,
-              })
-            : covalentApiKey
-            ? AVAXToolbox({
-                provider,
-                signer,
-                covalentApiKey,
-              })
-            : undefined;
+      const toolbox = await getToolbox({
+        session,
+        address,
+        chain,
+        walletconnect,
+        ethplorerApiKey,
+        covalentApiKey,
+      });
 
-        const address = await signer.getAddress();
-
-        addChain({
-          chain,
-          walletMethods: { ...toolbox, getAddress: () => address },
-          wallet: { address, balance: [], walletType: WalletOption.WALLETCONNECT },
-        });
-        return;
-      }
-      throw new Error('Chain is not supported');
+      addChain({
+        chain,
+        walletMethods: { ...toolbox, getAddress: () => address },
+        wallet: { address, balance: [], walletType: WalletOption.WALLETCONNECT },
+      });
+      return;
     });
 
     await Promise.all(promises);