+++ packages/wallets/ledger/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @thorswap-lib/ledger
 
+## 1.0.0-development.201
+
+### Patch Changes
+
+- Updated dependencies [[`3066000`](https://github.com/thorswap/SwapKit/commit/30660006cd0551dfb2be606290b0448bf0181d72)]:
+  - @thorswap-lib/types@1.0.0-development.94
+  - @thorswap-lib/helpers@0.0.0-development.94
+  - @thorswap-lib/swapkit-entities@1.0.0-development.102
+  - @thorswap-lib/toolbox-cosmos@1.0.0-development.122
+  - @thorswap-lib/toolbox-evm@1.0.0-development.138
+  - @thorswap-lib/toolbox-utxo@1.0.0-development.120
+
+## 1.0.0-development.200
+
+### Patch Changes
+
+- [#185](https://github.com/thorswap/SwapKit/pull/185) [`4e7b77a`](https://github.com/thorswap/SwapKit/commit/4e7b77adfd868ce26df05060f02dd2aab18dfa5d) Thanks [@chillios-ts](https://github.com/chillios-ts)! - Bump dependencies
+
+- Updated dependencies [[`4e7b77a`](https://github.com/thorswap/SwapKit/commit/4e7b77adfd868ce26df05060f02dd2aab18dfa5d)]:
+  - @thorswap-lib/swapkit-entities@1.0.0-development.101
+  - @thorswap-lib/toolbox-cosmos@1.0.0-development.121
+  - @thorswap-lib/toolbox-utxo@1.0.0-development.119
+  - @thorswap-lib/toolbox-evm@1.0.0-development.137
+  - @thorswap-lib/types@1.0.0-development.93
+  - @thorswap-lib/helpers@0.0.0-development.93
+
 ## 1.0.0-development.199
 
 ### Patch Changes